requirements.txt
langchain
langchain-community
langchain-openai
openai
python-telegram-bot
PyMuPDF
faiss-cpu
python-dotenv
psycopg2
tiktoken
pytest
pytest-asyncio
aiofiles
functools
python-json-logger